#80 Refactor : Reducer.js 파일 위치 변경
